Need to fix a out of bounds exception thats happening in the quick sort algorithm
i tuned every sort so that it takes in doubles so that its possible to compare the volumes. I dunno what we wanna do about areas, maybe a sub menu somewhere later on that asks what would u like to compare before the filling process starts. Anyway ill be working on this again shortly